Create AppFloatingActionButton.kt in core.presentation.components and correct FloatingActionButton in AddProductBottomSheet.kt
